(SSPP-255): Allow user to edit PDF text before upload
Created a new view where the user can now edit the scanned text before they create a PDF out of it. Moved the PDF creation into that new view. Fixed a bug inside the FilePickedView where it wouldn't reset the file when clicking the X button. Fixed a few UI issues as well. Also added a temporary fix to the refresh token not working on the physical device.