(SSPP-17/SSPP-227/SSPP-254/SSPP-163/SSPP-256): Alpha Pre-Release Version (staging)  (#14)
* (SSPP-228): Added classroom view UI and classroom models

Note: Data obtained from a json list with the same format as the classroom json output from postman

Also removed some sample testing images

* (SSPP-228): Minor changes to grid layout and background

* (SSPP-228/229/230): Added basic UI for join and create classrooms

SSPP-230: Updated UI
Also updated UI tool bar visuals in Notestorage and Noteupload

* (SSPP-228): Validation change

Added validation for users who are not in any classrooms

* (SSPP-228): Added classroom list UI integration (well it works) with the backend

Initial commit for a working class list UI integration with back end

* (SSPP-228): Removed some testing parameters from classroom list UI view

* (SSPP:229): Added backend integration to create classroom UI page

* (SSPP-230): Added join classroom UI integration with backend

* (SSPP-231): Added Search and top recommendation to classroom

-Included a clipboard-able class ID
-Updated MininoteView to properly display the note author

* (SSPP-231): Fixed a commented issue causing classrooms to not be displayed

* (SSPP-244): Implemented CD automated version tagging for client release versions (#13)

* (SSPP-244): Added CD workflow for git tagging

* (SSPP-244): Made version string semver

* (SSPP-244): Removed CI until later when automated testing plan is established

* (SSPP-228): Fixed class loading issue when there are no classes under the user

* (SSPP-231): Updated note integration with backend

* (SSPP-232): Added basic upload functionality

Fixed minor UI bugs

* SSPP-246: Update UI to include handwritten note option

Added a new button to the upload note view that will link to a new NoteScanView. Also updated the build version and added the required camera privacy settings to info.plist.

* (SSPP-250/SSPP-251/SSPP-252): Integrate VisionKit into new view for upload, Integrate PDFKit into the VisionKit view, Confirm that note upload still works with the new generated PDF files

Added the necessary PDFKit generation functionality to create PDF files out of the scanned images (one page per image). Then confirmed on my device that the upload still succeeds with the generated PDF files. Also adjusted the plist and build settings to allow the app to build on my physical device.

* (SSPP-253: Fix formatting of the generated PDFs)

Adjusted some code in the PDF generation process so the generated PDF files now look better.

* (SSPP-255): Allow user to edit PDF text before upload

Created a new view where the user can now edit the scanned text before they create a PDF out of it. Moved the PDF creation into that new view. Fixed a bug inside the FilePickedView where it wouldn't reset the file when clicking the X button. Fixed a few UI issues as well. Also added a temporary fix to the refresh token not working on the physical device.

* (SSPP-256): Create logout from profile view

Started creating the functionality to allow the user to log out. The logging out functionality works correctly so far, but it does not yet correctly return the user to the login screen after logging out.

* (SSPP-256): Create logout from profile view

Added a view model for the profile view and adjusted the authentication to navigate to the correct screen after hitting the logout button.

* (SSPP-255): Allow user to edit PDF text befored upload

Cleaned up the note editor view a bit and added a new icon for the handwritten note upload selection.

* (SSPP-17-new): Test merge SSPP-254 into SSPP-17

* (SSPP-17): Fix for LoginView navigation issue

Fixed an issue when logging in where clicking on some views would cause all teh views to shift upwards.

* (SSPP-178, SSPP-179): Fix UI shifting upwards during input

Fixed an issue on both the signup and upload view where opening the keyboard would cause the view to shift upwards and block access to certain inputs.

* (SSPP-163): Fix issue where erroneous note upload/creation acts successful on client

Fixed the bug when uploading a note where the upload would act successful even if it was not uploaded properly on the server. This happened whenever a server side error occurred. Also fixed a similar issue in the join classroom view.

Co-authored-by: Sheharyaar <maliksheharyaar@gmail.com>
Co-authored-by: Sheharyaar <38083700+maliksheharyaar@users.noreply.github.com>
Co-authored-by: Liam Stickney <liam.0820@hotmail.com>