(SSPP-283): Update Upload UI to include citation (#16)
Made updates to Neptune so BibtextCitation is not always nil when uploading. Made changes to the note upload view to include fields for citation author, title and year. These are all optional fields, and the citation will only be completed if all fields are entered. Also adjusted the UI so the screen looks less cluttered. We need to update the view model to validate these fields, specifically the year field. We also still need to make the citation visable on the note view UI.

Authored-by: Liam Stickney <liam.0820@hotmail.com>