(SSPP-172): User profile and management UI update
Attempted to get the profile view populated with some user data. Created a new function in the profile view model to obtain user information and tried to use it on the profile view to populate the text fields. However, XCode currently tries to display these values before the onAppear function is called in the ProfileView (which calls the function to get the user values from the view model), so all of the user values end up being nil and crash the app. I'm not sure if there's a better way to do this, but I tried to find a way to not force unwrap the values in the profile view, but I couldn't figure anything out. Currently this code builds, but crashes if you try to open the profile view.