(SSPP-290): Update Note Models
Neptune.swift, CloudNoteView.swift, NoteViewViewModel.swift and NoteSearchView.swift. For the note search view, I’ve hardcoded the rating displayed to just be [0,0,0,0,0] for now, as the ApiNoteResponse no longer contains a rating.

The ApiNoteRespone no longer has rating as an attribute, and its attribute ‘body’ has been renamed ‘noteAbstract’. As such, everywhere the ApiNoteReponse was used, these changes needed to be made.