(SSPP-172): User profile and management UI update
Updated the UI of the ProfileView a little bit and added some new functions to the ProfileViewModel to obtain information about the user’s classrooms and notes. Classrooms appear to work fine, but I am instantly logged out whenever I attempt to use the getUserNotes() in the NeptuneAPI. There appears to be some sort of issue in Neptune related to getting user notes that causes the user to be logged out instantly.