(SSPP-172): Liam implemented profile management views with data (#20)
* (SSPP-172): User profile and management UI update

Attempted to get the profile view populated with some user data. Created a new function in the profile view model to obtain user information and tried to use it on the profile view to populate the text fields. However, XCode currently tries to display these values before the onAppear function is called in the ProfileView (which calls the function to get the user values from the view model), so all of the user values end up being nil and crash the app. I'm not sure if there's a better way to do this, but I tried to find a way to not force unwrap the values in the profile view, but I couldn't figure anything out. Currently this code builds, but crashes if you try to open the profile view.

* (SSPP-172): User profile and management UI update

Updated the UI of the ProfileView a little bit and added some new functions to the ProfileViewModel to obtain information about the user’s classrooms and notes. Classrooms appear to work fine, but I am instantly logged out whenever I attempt to use the getUserNotes() in the NeptuneAPI. There appears to be some sort of issue in Neptune related to getting user notes that causes the user to be logged out instantly.

* (SSPP-172): Update profile view

Made a fix to the Neptune file and updated the profile view to include more information.

* (SSPP-172): User profile and management UI update

Updated the profile view to show all the information in some navigation views. Also moved the buttons for logging out and changing password into their own navigation view.

* (SSPP-305): Fixed bug in profile view not loading user data properly

Authored-by: Liam Stickney <liam.0820@hotmail.com>