(SSPP-284): Note citation during upload
Added a dependency to the SwiftyBibtex library as well as the required ANTLR library. Added the ability to use the view citation button in the note view. Clicking it will display the citation information provided during the original upload (author, title, year) in an alert, provided all the information was available. If any of the citation is missing, the alert will instead say no citation information is available. Tested this with a note with and a note without citation information.