(SSPP-284): Update Note View UI to include citation info
Added a sheet view to the note view UI that brings up three different kinds of citation styles using the information provided by the user during the note upload, assuming they were provided. The citations are available in text views that can be copied and pasted if the user long taps on them.