(SSPP-284): Implemented note view citation functionality (#23)
* (SSPP-284): Note citation during upload

Added a dependency to the SwiftyBibtex library as well as the required ANTLR library. Added the ability to use the view citation button in the note view. Clicking it will display the citation information provided during the original upload (author, title, year) in an alert, provided all the information was available. If any of the citation is missing, the alert will instead say no citation information is available. Tested this with a note with and a note without citation information.

* (SSPP-284): Update Note View UI to include citation info

Added a sheet view to the note view UI that brings up three different kinds of citation styles using the information provided by the user during the note upload, assuming they were provided. The citations are available in text views that can be copied and pasted if the user long taps on them.

* (SSPP-284): Added some UI enhancements to the Note + Citations views

* (SSPP-284): Implemented UI enhancements for the note detail view and citations

Co-authored-by: Liam Stickney <liam.0820@hotmail.com>