(SSPP-309): Change Password UI Update
Added the ability on the client to change the user's password. This is available through the profile view. The user can click on the button and it will allow them to enter in both their new and old passwords. The change password will only work if the old password is valid and the passwords do not match. If it is successful, the app will log the user out and prompt them to sign in again with their new password.