(SSPP-309): Implemented client-side password change for the user profile (#27)
* (SSPP-299): Added some PDFKit code to work more on later

* (SSPP-299): View note PDF from remote CDN

* (SSPP-309): Change Password UI Update

Added the ability on the client to change the user's password. This is available through the profile view. The user can click on the button and it will allow them to enter in both their new and old passwords. The change password will only work if the old password is valid and the passwords do not match. If it is successful, the app will log the user out and prompt them to sign in again with their new password.

Co-authored-by: Liam Stickney <liam.0820@hotmail.com>