(v1.2.0): Final Capstone Release (#31)
* (SSPP-306): Implemented auth flow for class detailed view + fixed some bugs with async callbacks done within class create, class join and class viewing (#22)

* (SSPP-284): Implemented note view citation functionality (#23)

* (SSPP-284): Note citation during upload

Added a dependency to the SwiftyBibtex library as well as the required ANTLR library. Added the ability to use the view citation button in the note view. Clicking it will display the citation information provided during the original upload (author, title, year) in an alert, provided all the information was available. If any of the citation is missing, the alert will instead say no citation information is available. Tested this with a note with and a note without citation information.

* (SSPP-284): Update Note View UI to include citation info

Added a sheet view to the note view UI that brings up three different kinds of citation styles using the information provided by the user during the note upload, assuming they were provided. The citations are available in text views that can be copied and pasted if the user long taps on them.

* (SSPP-284): Added some UI enhancements to the Note + Citations views

* (SSPP-284): Implemented UI enhancements for the note detail view and citations

Co-authored-by: Liam Stickney <liam.0820@hotmail.com>

* (SSPP-124): Added Personal Note Storage View + Ability to Delete Notes (#24)

* (SSPP-124): Updated personal storage UI to load user personal notes and to be directed to them when tapped on

Note: Delete functionality is temporarily disabled

* (SSPP-268): Added note deletion from personal storage and update rowitemViews for both classroom dashboard and personal storage

Co-authored-by: Sheharyaar <38083700+maliksheharyaar@users.noreply.github.com>

* (SSPP-299): Implemented note file view from remote CDN using PDFKit (#25)

* (SSPP-299): Added some PDFKit code to work more on later

* (SSPP-299): View note PDF from remote CDN

* (SSPP-286): Implemented Note Rating Capabilities (#26)

* (SSPP-286): Added disable feature to rating UI as to stop it from being intractable on views it shouldnt. Also updated buttons on login and signup views

* (SSPP-286): Added partial update of note rating, still need to display the average note rating properly

Just in case commit if my comp decided to go poof

* (SSPP-286): Added display of average note in personal storage and classroomDetailView. There is a possibility that it spams docker with requests and gets locked out until the docker is restarted

POTENTIAL BUG: May or may not spam requests when updating rating and results in crashing the docker. Docker then needs to be restarted

Co-authored-by: Sheharyaar <38083700+maliksheharyaar@users.noreply.github.com>

* (SSPP-309): Implemented client-side password change for the user profile (#27)

* (SSPP-299): Added some PDFKit code to work more on later

* (SSPP-299): View note PDF from remote CDN

* (SSPP-309): Change Password UI Update

Added the ability on the client to change the user's password. This is available through the profile view. The user can click on the button and it will allow them to enter in both their new and old passwords. The change password will only work if the old password is valid and the passwords do not match. If it is successful, the app will log the user out and prompt them to sign in again with their new password.

Co-authored-by: Liam Stickney <liam.0820@hotmail.com>

* (SSPP-182): Added ToS + Fix for rating issue + Updating styling for detail note view (#28)

* (SSPP-182/319/316): Added TOS to Signup view, Fixed issue with inaccurate rating values, Added rating counter

SSPP-182: Added a Terms and Conditions alert that must be viewed and accepted by the user to make a StudySnap account

SSPP-316: Added a note rating counter to row items to display the total number of ratings a note has

SSPP-319: Fixed a issue regarding inaccurate rating where it was using the average rating as the individual rating value when the user updated rating by going back and forth from different views

* (SSPP-182): Fixed a minor spelling error

* (SSPP-182): Updated UI for Detailed note Get full note icon

* (release): updated version string for v1.1.1 release

Co-authored-by: Sheharyaar <38083700+maliksheharyaar@users.noreply.github.com>

* (SSPP-307): Implemented async loading indicators where needed (#29)

* (SSPP-307): Added loading indicators where needed + UI enhancements

* (SSPP-307): Updated compose environment versions

* (SSPP-307): Ensured any errors interrupted loading on join and create classrooms

* (SSPP-313): Edit note features along with a number of small UI fixes (#30)

* (SSPP-313): Added edit note functionality for title and short description, and updated UI for Detail class view, Personal Storage, Profile view

SSPP-313: Edit functionality is handled using a custom halfsheet view and allows user to change either title or short description.

If user only changes one field then the other filed uses the already exiting value.

Non note authors cannot make changes to someone else's note

* (SSPP-313): Added a visible grab icon for the halfsheet view and updated color asset

Also added dismissal of halfsheet view if a tap is detected on outside of the sheet boundary

Note: If any reason this commit may result in issues, then look at the comments in DetailNoteView and use the default sheet modifier code instead on the custom halfsheet modifier

* (develop -> SSPP-313): Created new minor version and caught up SSPP-313

* (SSPP-313): Minor UI adjustments

* (SSPP-313): added keywords as taglist in upload view

* (SSPP-313): added keywords ui enhancment to note upload view

* (SSPP-313):Added refresh to row items when navigating back to the view from the detail note view

* (SSPP-313): Added a refresh icon on personal note view

Co-authored-by: Sheharyaar <38083700+maliksheharyaar@users.noreply.github.com>

Co-authored-by: Liam Stickney <liam.0820@hotmail.com>
Co-authored-by: Sheharyaar <38083700+maliksheharyaar@users.noreply.github.com>